feat:update payment section in registration page
