+++ src/app/events/page.jsx
@@ -1,9 +1,5 @@
 'use client';
-// import {
-//   GL_descriptionItems,
-//   GL_previewItems,
-// } from '@/config/content/EventsPage/GuestLecture/Data';
-// import { Heading3 } from '@/components/shared/Typography/Headings';
+
 import { EventsContainer, Heading3 } from './page.style';
 import MainCarousel from '@/components/EventsPage/index';
 import {
@@ -14,7 +10,6 @@ import {
   Exhibition_descriptionItems,
   Exhibition_previewItems,
 } from '@/config/content/EventsPage/Exhibition/Data';
-// import { DTS_previewItems, DTS_descriptionItems } from '@/config/content/EventsPage/DTS_Shows/Data';
 import { ME_descriptionItems, ME_previewItems } from '@/config/content/EventsPage/MainEvents/Data';
 
 import {
@@ -23,8 +18,6 @@ import {
 } from '@/config/content/EventsPage/FunEvents/Data';
 import { ComingSoon } from '@/components/EventsPage/ComingSoon';
 import { GalleryWrapper } from '@/components/EventsPage/Gallery/CardWrapper';
-// import { GalleryWrapper, Gallerywrapper } from '@/components/EventsPage/Gallery/CardWrapper';
-// import { GalleryWrapper } from '@/components/EventsPage/Gallery/CardWrapper';
 
 export default function Page() {
   return (
@@ -45,10 +38,8 @@ export default function Page() {
           descriptionItems={Exhibition_descriptionItems}
         />
         <Heading3>SHOWS AT DTS</Heading3>
-        {/* <MainCarousel previewItems={DTS_previewItems} descriptionItems={GL_descriptionItems} /> */}
         <ComingSoon />
         <Heading3>GUEST LECTURES</Heading3>
-        {/* <MainCarousel previewItems={GL_previewItems} descriptionItems={GL_descriptionItems} /> */}
         <ComingSoon />
         <Heading3>GALLERY</Heading3>
         <GalleryWrapper />
