+++ src/components/EventsPage/Carousel/PreviewCarousel.styles.jsx
