+++ src/components/EventsPage/Carousel/swiper.css
@@ -3,6 +3,12 @@
   height: 100%;
   padding-top: 90px;
 }
+.mySwiper2 {
+  width: 100%;
+  height: auto;
+  overflow: hidden;
+  padding: 20px;
+}
 .mySwiper3 {
   width: 100%;
   height: 100%;
@@ -13,13 +19,6 @@
   width: 100%;
   height: 100%;
   padding-top: 90px;
-}
-
-.mySwiper2 {
-  width: 100%;
-  height: auto;
-  overflow: hidden;
-  padding: 20px;
 }
 
 .gallery-swiper {
