+++ src/components/EventsPage/Shared/ArrowButton.jsx
@@ -1,7 +1,7 @@
 import tw from 'twin.macro';
 import styled from 'styled-components';
 
-const commonStyles = tw`absolute top-1/4 transform -translate-y-1/2 bg-cover bg-no-repeat w-11 h-16 cursor-pointer transition-all duration-200 ease-in-out`;
+const commonStyles = tw`absolute top-1/4 transform -translate-y-1/2 bg-cover bg-no-repeat w-11 h-16 cursor-pointer transition-all duration-200 ease-in-out hidden sm:block`;
 
 const LeftArrowButton = styled.button`
   background-image: url('https://res.cloudinary.com/dhv234qct/image/upload/v1728883189/Inno2k24/o9yetaxkolszugpanrje.svg');
