+++ src/components/EventsSection/eventCardComponents/Card.jsx
@@ -1,8 +1,7 @@
 import Image from 'next/image';
 import Link from 'next/link';
 
-export const EventCard = ({ label, isTranslatedUp }) => {
-  console.log(label);
+export const EventCard = ({ label }) => {
   return (
     <>
       <Link href={'/events'}>
